Object detection/Obj_detect_api/requirements.txt
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Cors==3.0.10
gunicorn
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
opencv-python==3.4.17.63
opencv-python--headless
Pillow==9.1.0
six==1.16.0
Werkzeug==2.1.1
zipp==3.8.0
